+++ CHANGELOG.md
@@ -97,3 +97,10 @@ Now plugin always must receive `params` object
 * Update dependencies
 * Update logic of work of `RatingChangedListener`: now by default first update for a long time will
 be displayed urgent, but all next updates in one time will be delayed with `debounceDelay`
+
+### 1.1.1
+
+* Update version of bot `1.2.0` - `1.2.6`
+* Rewrite system of registering for messages and posts likes groups
+* Databases channels now have capacity `Channel.CONFLATED`
+* Added command `/attachSeparatedLikes`